cros: Actually implement VbExEcGetExpectedRWHash.
The new implementation returns the hash which has been put in the device tree
by cros_bundle_firmware. The amount of data that's loaded from flash is
however much has been signed by the vblock. To avoid loading the EC
prematurely we simply drop it from the list of signed blobs. When EC software
sync wants a copy of the RW to hash and send to the EC because it needs to do
an update, there was already a redundant load to get the image out of flash.
That load is no longer redundant.

BUG=chrome-os-partner:17606
TEST=Built on Link and measured a 170ms speed up on RW boots. Built on Snow
and measured boot times which erratic but comparable before and after this
change. Ran the EC software sync FAFT test on Link.
BRANCH=None

Change-Id: I3d78550180a772388f2ae028df24f345a1263a57
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/46727
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>