spring: low battery mode
When the system is started with low battery charge and the power supply
is not providing enough current, just display a low battery screen, then
shutdown.

The proper power management for the low battery screen is not
implemented in this CL.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:17232 chrome-os-partner:17268
TEST=on Spring, fake battery level with the EC and check system
behavior.

Change-Id: I268b1e3e889eb344091941a918cc752bc36091ce
Reviewed-on: https://gerrit.chromium.org/gerrit/45636
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>