exynos: add a third GPIO for board revision on Spring
GPD0_2 has been added on Spring EVT to have a third (trinary) bit for
the board SKU/revision number.
This pin was floating (which is decoded as 0) on previous revisions of
the board, so this should be backward compatible.
Starting DVT2, this bit will set (rev 9+).

Also increase CONFIG_BOARD_REV_GPIO_COUNT which the maximum number of
GPIOs used for revision encoding rather than the actual number,
so this should not be an issue on boards with only 2 GPIOs.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:17571 chrome-os-partner:19926
TEST=boot on Spring EVT and see the board revision is still 5.

Change-Id: Ie1d9ba8e3787b9d815e347635db72afa8d4d0cef
Reviewed-on: https://gerrit.chromium.org/gerrit/57969
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>