spring: reduce backlight for low battery screen
Set the backlight PWM to 7% when displaying the low battery screen,
this should be acceptable to see a high contrast battery pattern.

With 7% backlight, it's currently consuming about 5.6 W at the low
battery screen. Given that the screen is displayed for 10 seconds,
we consume about 16 mWh (on a 30 Wh battery) or about 0.05% of the
capacity.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:17232
TEST=on Spring, run with an EC returning a fake battery level/charger
presence, and read the power consumed from the smart battery.

Change-Id: Ie4516512cc89603b40705fb419332e1e760b90bf
Reviewed-on: https://gerrit.chromium.org/gerrit/45637
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>