input: refine ghost keys detection
If the fourth key of a ghost key combination is not used in the matrix,
there is no possible ghosting issue. So we don't want to drop the keys
in that situation.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:18354
TEST=on Spring with DEBUG activated in key_matrix.c, press various
interesting key combinations and see the result on the firmware serial
console.

Change-Id: I826f5d270deee8cac3860c3ef7b9d656dbaa420f
Reviewed-on: https://gerrit.chromium.org/gerrit/46142
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>