vboot: do not reboot the platform when sound init fails
When the sound initialization is failing, we don't really want to panic
and cold reboot the firmware as this is not really user friendly.
Let just print the failure and continue.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:18690
TEST=on Spring EVT with unsupported codec, wait 30s at dev screen and
see the machine booting from the eMMC instead of rebooting to the dev
screen.

Change-Id: I496ef1ccc0f04e84c73b70beeda57715faa1a629
Reviewed-on: https://gerrit.chromium.org/gerrit/47876
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>