DO NOT SUBMIT: exynos: hack to boot snow with simplefb
This is a local hack for nv-u-boot to pass in simplefb settings
and keeping the panel alive when booting the kernel. This way the
kernel can use it without having drivers for DP, etc.

This is in other words just a temporary hack until sufficient kernel
support lands to bring up graphics natively there, to build a nv-u-boot
for chainloading with.

Change-Id: I30bf53491ae1973715d596d0758335ad1c83f1ca
Signed-off-by: Olof Johansson <olofj@chromium.org>
Signed-off-by: Andreas Färber <andreas.faerber@web.de>