elog: Add configuration information to the device tree.
The full size of the log, the size to shrink it by when it's full, and a few
other parameters are stored in the device tree to be retrieved by the event
log code at run time.

BUG=chrome-os-partner:16132
TEST=With this and other changes, booted on spring and used mosys eventlog
list to look at the events in the log.
BRANCH=spring

Change-Id: I757c89e0a878bea7174a9268c7a50f95ccd9a3a1
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/49873
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>