elog: Make it possible to build the event log and rtc code into SPL
Only the SPL part of firmware participates in resume, so it's where we need to
record a resume event. This change makes it possible to build the event log
and RTC code into the SPL.

To get the RTC code to link properly in the SPL, a variable needs to be forced
into the data section instead of the bss.

Also, in the SPL, we can't copy the contents of flash into memory like the
elog code normally would. Instead, we have to parse the log manually. Some of
the event log code is still useful, though, specifically the code that puts
together an event. Only that code is included when building the SPL.

To make it easier to support the event log with custom code in the spl, some
data structures are moved from the internal header file to the external
header.

BUG=chrome-os-partner:16132
TEST=Built and booted on spring.
BRANCH=spring

Change-Id: I99b7bab6f95ffbb92aa5f3a8d54e4007b077f140
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/50147
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>