elog: Steal some space from ramoops for the event log panic buffer
This range of memory will be used to tuck away events during a panic that
the firmware is supposed to record later. We steal some space from ramoops
so that the remaining memory areas are still aligned with large enough
granularity that the kernel can handle them.

BUG=chrome-os-partner:16132
TEST=With this and other changes, built and booted on Spring. Verified that
panic events were passed back from the kernel correctly and recorded in the
event log.
BRANCH=spring

Change-Id: I1a2c2b461711968b8968b49d14ed2e456f1926e3
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/51449
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>