elog: Copy panic events from the kernel into the event log in flash
This change adds a function which scans the panic event buffer and copies
any event it finds which is legally formed into the event log in flash. As
soon as it runs out of events, it stops looking through the buffer. Once any
events are copied over, the buffer is memset to zero.

When panic events are recorded in memory by the kernel, we want to add them
to the log directly without updating their timestamp, so this change adds a
function to do that.

BUG=chrome-os-partner:16132
TEST=Built and booted on Spring. Used prototyped support in the kernel to add
panic events to the buffer and saw that the events were added to the log once
the system had rebooted. Verified that boot and resume events were still
added.
BRANCH=spring

Change-Id: Ib75a639b8098731719e3d7dbc4ad06f75939962d
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/51451
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>