elog: Pass event log info to the kernel, and trim out the panic buffer
Pass the location of the event log panic buffer and the offset of the event
log itself to the kernel through its device tree. Also, trim the panic buffer
range out of the /memory/reg property to keep the kernel from using it for
something else.

BUG=chrome-os-partner:16132
TEST=Built and booted on Spring. With prototyped panic code, verified that a
panic event was stored in the right place, and was picked up by the firmware
and recorded. Looked at the device tree in the booted system and verified
that the expected properties where present in the right places.
BRANCH=spring

Change-Id: I08ecad1b1f90efcdac41d814524b27d21b7f0e9b
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/51452
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>