rtc: Add rtc type to the machine params
The eventlog code needs the RTC in SPL, and the RTC needs to be able to tell
if it should run.

BUG=chrome-os-partner:16132
TEST=With this and other changes, built and booted on Spring. Verified that
a wake event is recorded with a time stamp, although the rtc driver can't talk
to the RTC over I2C after a resume for some reason.
BRANCH=spring

Change-Id: Ic4e250018328096059dc1221e1ae7cb0c2207a5f
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/51448
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>