elog: If reading from the RTC fails, use a fixed timestamp
When reading from the RTC fails, there's no guarantee what's in the timestamp
when the reading function returns. We should set it to something explicit so
we can be confident whether a timestamp is valid.

BUG=chrome-os-partner:16132
TEST=Built and booted on Spring. Suspended and resumed and used mosys eventlog
list to look at the log and verify that the event was recorded as expected.
BRANCH=spring

Change-Id: I10d4a177d5b0ad2dff78e5695b68ebd6d713ef26
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/58502
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>