elog: Use a new, generic wake event for ARM
The existing wake event is for use with ACPI and isn't entirely appropriate
for ARM. This change adds a new event which has no data payload and uses it
for resume for exynos.

BUG=chrome-os-partner:16132
TEST=Built and booted on Spring. Did a suspend/resume cycle and used mosys
eventlog list to check that the event was recorded correctly.
BRANCH=spring

Change-Id: I172cf1ad17ce017e31f688ed280a1d234f065392
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/58503
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>