eventlog: Move the panic event buffer away from ramoops.
The interaction of the panic event buffer and the ramoops area has been
problematic. If the two together don't remove a chunk of memory from the map
which is 1MB aligned at all boundaries, then the kernel will hang in early
startup. To avoid that problem, the panic event buffer is being moved to not
overlap with the current position of the ramoops area and so that it can be
padded out to 1MB without interfering with anything else.

BUG=chrome-os-partner:16132
TEST=Built and booted on spring. With an enabling kernel change, caused a
panic and saw that the panic event was recorded in the event log after the
reboot.
BRANCH=spring

Change-Id: Idf48975da0700a25b39b761ac574b441a9b766cc
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/62330
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>