ARM64: rockchip: Add initial support for RK3368 based GeekBox
The RK3368 is an octa-core Cortex-A53 SoC from Rockchip.

The GeekBox is a TV box from GeekBuying, based on an MXM3 module.
The module can be used with base boards such as the GeekBox Landingship.

This adds basic support to chain-load U-Boot from Rockchip's miniloader.

  $ ./lollipop_u-boot/tools/loaderimage --pack u-boot.bin u-boot.img
  # ./utils/upgrade_tool di uboot u-boot.img

Implemented is the serial console, but no boot media drivers yet.

Note that flashing the resulting U-Boot will not allow you to enter the
rockusb mode any more via "Update" button. Instead, you will need to
short two pins on the bottom of the module to enter MaskRom mode and
re-flash the loader:

  # ./utils/upgrade_tool ul ./lollipop_u-boot/RK3368MiniLoaderAll_V2.40.bin
  # ./utils/upgrade_tool di uboot u-boot.img

Signed-off-by: Andreas Färber <afaerber@suse.de>